==> processing/01_merging_files.ipynb <==
# Merging Thousands of ```.txt()``` Files into a ```.parquet``` File
### Merging all InfoUSA files and saving them as a ```.parquet``` file

### Import statements

import pyarrow
import re
import os
import pandas as pd

This will stop warnings from showing up.

import warnings
warnings.filterwarnings('ignore')

### Setting ```DATA_DIR```
In order to read in files from this repository, we must set ```DATA_DIR``` to be the data folder within this repository. This requires ```os.getcwd()``` to return the path to the processing notebook of this repository, so ```xxx/codeplus-celine-dcc-package/procesing```, where ```xxx``` is the path to where you cloned this repository. If it is not, use ```os.chdir(path)``` to change the current working directory to ```xxx/codeplus-celine-dcc-package/procesing``` before getting the current working directory in ```DATA_DIR = os.getcwd()```, where ```path``` is ```xxx/codeplus-celine-dcc-package/procesing```.

DATA_DIR = os.getcwd()
DATA_DIR = DATA_DIR.replace('processing', 'data')
DATA_DIR

### Merging files
We initially attempted to merge all files in one go and save the combined file as a ```.csv()``` file. Although we were able to do this, reading in the combined ```.csv()``` file would not work- our kernel crashed every time we attempted it, even using 38 CPU cores and 208GB of RAM. 

Therefore, we decided to use pyrarrow as an engine when reading in each of the ```.txt``` files and save the combined file as a ```.parquet()``` file. Pyarrow allows for multithreading, leading to a more efficient use of CPU cores, and Parquet files are optimized for storing large datasets. This is because Parquet files are column-based while CSV files are row-based, meaning it can only focus on the data relevant to the user's query. Like this, it only loads into memory the columns of the data that are being used. Parquet files can also be read in and exported using pandas, which is an added bonus when working with pandas dataframes.

Still, we ran into memory issues merging all files in one go- so we broke up our data into four sections, merged the files in each of these sections, dropped columns not relevant to our work, then combined the four sections for one totally combined dataframe.

Here, we would merge zip codes in the range 00000 - 25000. The regular expression sifts through the list of all files in the InfoUSA folder and creates a list of only the files with zip codes in that range. However, using the synthetic InfoUSA data we created in processing notebook **synthetic_infousa**, we do not have to break it up into sections-- we only included 10 files, totalling 1 million observations. Therefore, this process only needs to be done once. We have included the original code for merging all InfoUSA files, though, below.

os.chdir(DATA_DIR + '/source_files/infousa_files')
directory = os.listdir('./')
print(directory)

Then, we use the pandas ```.concat()``` function to concatenate every file in the list of files created above.

%%time
#combine all files in the list
df = pd.concat([pd.read_csv(f, sep = '\t', dtype = {'census_county_2010': 'str', 'census_state_2010': 'str'}) 
                for f in directory if f[0:9] == 'Household'])
df

Filter for only the columns relevant to our future work:

%%time
df['county_fips'] = df['census_state_2010'] + df['census_county_2010']

df = df[['ZIP', 'county_fips', 'STATE', 'ChildrenHHCount', 'children_ind', 'head_hh_age_code', 'GE_LATITUDE_2010', 'GE_LONGITUDE_2010']]

Filter for only zipcodes not in Alaska, Hawaii, Puerto Rico, and the Virgin Islands, as there are no storage tanks in the AST dataset from these areas:

%%time
df = df[((df['STATE'] != 'AK') & (df['STATE'] != 'HI') & (df['STATE'] != 'PR') & (df['STATE'] != 'VI'))]

#### Renaming columns
We rename the columns in our dataset for standardization purposes.

df.rename(columns = {'ZIP': 'zip', 'STATE': 'state', 'ChildrenHHCount': 'child_num', 
                           'children_ind': 'has_child', 'head_hh_age_code': 'age_code', 'GE_LATITUDE_2010': 'lat_h_4326', 
                            'GE_LONGITUDE_2010': 'lon_h_4326'}, inplace = True)
df

#### Transforming household latitude and longitude coordinates from EPSG 4326 to EPSG 3857
A lot of our visualizations need coordinates in EPSG 3857, however these coordinates are in EPSG 4326. Therefore, we use the pyproj interface, which allows us to use the PROJ coordinate transformation software to transform our EPSG 4326 coordinates to EPSG 3857. This creates two new columns in our original dataset with the transformed coordinates.

from pyproj import Proj, Transformer

# Apply transformation
transform_4326_to_3857 = Transformer.from_crs('epsg:4326', 'epsg:3857')
df['lat_h_3857'], df['lon_h_3857'] = transform_4326_to_3857.transform(
                                                df['lat_h_4326'], df['lon_h_4326'])

df

#### Exporing final dataframe
Finally, we export this dataframe as a ```.parquet()``` file for further use.

%%time
df.to_parquet(DATA_DIR + '/infousa_merged.parquet')

f = pd.read_parquet(DATA_DIR + '/infousa_merged.parquet')
f

### Merging InfoUSA files - original code

### Merging 00000-25000
Using the same steps as above, but this time using regular expression to filter for zipcodes within the range stipulated.

os.chdir("/hpc/group/codeplus22-vis/infousa_copy/2020")
directory = os.listdir("/hpc/group/codeplus22-vis/infousa_copy/2020")
r = re.compile("Household_Ethnicity_zip_(0{4}[0-9]|0{3}[1-9][0-9]|00[1-9][0-9]{2}|0[1-9][0-9]{3}|1[0-9]{4}|2[0-4][0-9]{3}|25000)_year_2020.txt")
newlist = list(filter(r.match, directory)) # Read Note
# print(newlist)
len(newlist)

Then, we use the pandas ```.concat()``` function to concatenate every file in the list of files created above.

%%time
#combine all files in the list
zip_00_25 = pd.concat([pd.read_csv(f, sep = '\t', 
                                   dtype = {'census_state_2010': 'str', 'census_county_2010': 'str'}) for f in newlist])
zip_00_25.head()

Filter for only the columns relevant to our future work:

%%time
zip_00_25['county_fips'] = zip_00_25['census_state_2010'] + zip_00_25['census_county_2010']

zip_00_25 = zip_00_25[['ZIP', 'county_fips', 'STATE', 'ChildrenHHCount', 'children_ind', 
                       'head_hh_age_code', 'GE_LATITUDE_2010', 'GE_LONGITUDE_2010']]

zip_00_25

Filter for only zipcodes not in Alaska, Hawaii, Puerto Rico, and the Virgin Islands, as there are no storage tanks in the AST dataset from these areas:

%%time
zip_00_25 = zip_00_25[((zip_00_25['STATE'] != 'AK') & (zip_00_25['STATE'] != 'HI') & (zip_00_25['STATE'] != 'PR') & (zip_00_25['STATE'] != 'VI'))]

zip_00_25

Write out the dataframe as a ```.parquet()``` file:

%%time
zip_00_25.to_parquet('/hpc/group/codeplus22-vis/infousa_copy/zip_00_25_final_fixed.parquet')

### Merging 25000-50000
We use the same steps as Merging 00000 - 25000 section.

os.chdir("/hpc/group/codeplus22-vis/infousa_copy/2020")
directory = os.listdir("/hpc/group/codeplus22-vis/infousa_copy/2020")
r = re.compile("Household_Ethnicity_zip_(2500[0-9]|250[1-9][0-9]|25[1-9][0-9]{2}|2[6-9][0-9]{3}|[34][0-9]{4}|50000)_year_2020.txt")
newlist = list(filter(r.match, directory)) # Read Note
# print(newlist)
len(newlist)

zip_25_50 = pd.concat([pd.read_csv(f, sep = '\t', 
                                   dtype = {'census_state_2010': 'str', 'census_county_2010': 'str'}) for f in newlist])
zip_25_50.head()

%%time
zip_25_50['county_fips'] = zip_25_50['census_state_2010'] + zip_25_50['census_county_2010']
zip_25_50 = zip_25_50[['ZIP', 'county_fips', 'STATE', 'ChildrenHHCount', 'children_ind', 'head_hh_age_code', 'GE_LATITUDE_2010', 'GE_LONGITUDE_2010']]

%%time
zip_25_50 = zip_25_50[((zip_25_50['STATE'] != 'AK') & (zip_25_50['STATE'] != 'HI') & (zip_25_50['STATE'] != 'PR') & (zip_25_50['STATE'] != 'VI'))]
zip_25_50

%%time
zip_25_50.to_parquet('/hpc/group/codeplus22-vis/infousa_copy/zip_25_50_final_fixed.parquet')

### Merging 50000-75000
We use the same steps as Merging 00000 - 25000 section.

os.chdir("/hpc/group/codeplus22-vis/infousa_copy/2020")
directory = os.listdir("/hpc/group/codeplus22-vis/infousa_copy/2020")
r = re.compile("Household_Ethnicity_zip_(5000[0-9]|500[1-9][0-9]|50[1-9][0-9]{2}|5[1-9][0-9]{3}|6[0-9]{4}|7[0-4][0-9]{3}|75000)_year_2020.txt")
newlist = list(filter(r.match, directory)) # Read Note
# print(newlist)
len(newlist)

%%time
#combine all files in the list
zip_50_75 = pd.concat([pd.read_csv(f, sep = '\t', dtype = {'census_state_2010': 'str', 'census_county_2010': 'str'}) for f in newlist])
zip_50_75.head()

%%time
zip_50_75['county_fips'] = zip_50_75['census_state_2010'] + zip_50_75['census_county_2010']
zip_50_75 = zip_50_75[['ZIP', 'county_fips', 'STATE', 'ChildrenHHCount', 'children_ind', 'head_hh_age_code', 'GE_LATITUDE_2010', 'GE_LONGITUDE_2010']]

%%time
zip_50_75 = zip_50_75[((zip_50_75['STATE'] != 'AK') & (zip_50_75['STATE'] != 'HI') & (zip_50_75['STATE'] != 'PR') & (zip_50_75['STATE'] != 'VI'))]
zip_50_75

%%time
zip_50_75.to_parquet('/hpc/group/codeplus22-vis/infousa_copy/zip_50_75_final_fixed.parquet')

### Merging 75000-99000
We use the same steps as Merging 00000 - 25000 section.

os.chdir("/hpc/group/codeplus22-vis/infousa_copy/2020")
directory = os.listdir("/hpc/group/codeplus22-vis/infousa_copy/2020")
r = re.compile("Household_Ethnicity_zip_(7500[0-9]|750[1-9][0-9]|75[1-9][0-9]{2}|7[6-9][0-9]{3}|[89][0-9]{4})_year_2020.txt")
newlist = list(filter(r.match, directory)) # Read Note
# print(newlist)
len(newlist)

%%time
#combine all files in the list
zip_75_99 = pd.concat([pd.read_csv(f, sep = '\t', dtype = {'census_state_2010': 'str', 'census_county_2010': 'str'}) for f in newlist])
zip_75_99.head()

%%time
zip_75_99['county_fips'] = zip_75_99['census_state_2010'] + zip_75_99['census_county_2010']
zip_75_99 = zip_75_99[['ZIP', 'county_fips', 'STATE', 'ChildrenHHCount', 'children_ind', 'head_hh_age_code', 'GE_LATITUDE_2010', 'GE_LONGITUDE_2010']]

%%time
zip_75_99 = zip_75_99[((zip_75_99['STATE'] != 'AK') & (zip_75_99['STATE'] != 'HI') & (zip_75_99['STATE'] != 'PR') & (zip_75_99['STATE'] != 'VI'))]

zip_75_99

%%time
zip_75_99.to_parquet('/hpc/group/codeplus22-vis/infousa_copy/zip_75_99_final_fixed.parquet')

### Merging 00000 - 50000
Now that we merged each of our four sections, we will merge these four sections with eachother using pandas' ```.append()``` function.

%%time
zip_00_50 = pd.read_parquet('/hpc/group/codeplus22-vis/infousa_copy/zip_00_25_final_fixed.parquet').append(pd.read_parquet('/hpc/group/codeplus22-vis/infousa_copy/zip_25_50_final_fixed.parquet'), ignore_index = True)
zip_00_50

%%time
zip_00_50.to_parquet('/hpc/group/codeplus22-vis/infousa_copy/zip_00_50_final_fixed.parquet')

### Merging 50000 - 99000
We use the same steps as Merging 00000 - 50000 section.

%%time
zip_50_99 = pd.read_parquet('/hpc/group/codeplus22-vis/infousa_copy/zip_50_75_final_fixed.parquet').append(pd.read_parquet('/hpc/group/codeplus22-vis/infousa_copy/zip_75_99_final_fixed.parquet'), ignore_index = True)
zip_50_99

%%time
zip_50_99.to_parquet('/hpc/group/codeplus22-vis/infousa_copy/zip_50_99_final_fixed.parquet')

### Merging 00-99
We use the same steps as Merging 00000 - 25000 section, but this time merge 00000 - 50000 and 50000 - 99999 to get our final, totally combined dataframe.

%%time
zip_00_99 = pd.read_parquet('/hpc/group/codeplus22-vis/infousa_copy/zip_00_50_final_fixed.parquet').append(pd.read_parquet('/hpc/group/codeplus22-vis/infousa_copy/zip_50_99_final_fixed.parquet'), ignore_index = True)
zip_00_99

### Renaming columns
We rename the columns in our dataset for standardization purposes.

zip_00_99.rename(columns = {'ZIP': 'zip', 'census_county_2010': 'county', 'STATE': 'state', 'ChildrenHHCount': 'child_num', 
                           'children_ind': 'has_child', 'head_hh_age_code': 'age_code', 'GE_LATITUDE_2010': 'lat_h_4326', 
                            'GE_LONGITUDE_2010': 'lon_h_4326'}, inplace = True)
zip_00_99

### Transforming household latitude and longitude coordinates from EPSG 4326 to EPSG 3857
A lot of our visualizations need coordinates in EPSG 3857, however these coordinates are in EPSG 4326. Therefore, we use the pyproj interface, which allows us to use the PROJ coordinate transformation software to transform our EPSG 4326 coordinates to EPSG 3857. This creates two new columns in our original dataset with the transformed coordinates.

from pyproj import Proj, Transformer

# Apply transformation
transform_4326_to_3857 = Transformer.from_crs('epsg:4326', 'epsg:3857')
zip_00_99['lat_h_3857'], zip_00_99['lon_h_3857'] = transform_4326_to_3857.transform(
                                                zip_00_99['lat_h_4326'], zip_00_99['lon_h_4326'])

zip_00_99

### Exporing final dataframe
Finally, we export this dataframe as a ```.parquet()``` file for further use.

%%time
zip_00_99.to_parquet('/hpc/group/codeplus22-vis/infousa_copy/zip_00_99_final_fixed.parquet')
