==> processing/03_children_per_county.ipynb <==
## Children Household Count Per County File Processing

In this file, we will be making a dataframe that includes the number of children in all of the households in each county of the US.

### Import statements

import pandas as pd
import geopandas as gpd
import os

### Setting ```DATA_DIR```
In order to read in files from this repository, we must set ```DATA_DIR``` to be the data folder within this repository. This requires ```os.getcwd()``` to return the path to the processing notebook of this repository, so ```xxx/codeplus-celine-dcc-package/procesing```, where ```xxx``` is the path to where you cloned this repository. If it is not, use ```os.chdir(path)``` to change the current working directory to ```xxx/codeplus-celine-dcc-package/procesing``` before getting the current working directory in ```DATA_DIR = os.getcwd()```, where ```path``` is ```xxx/codeplus-celine-dcc-package/procesing```.

DATA_DIR = os.getcwd()
DATA_DIR = DATA_DIR.replace('processing', 'data')
DATA_DIR

### Importing county level shape file

This shapefile of US counties was taken from the United States Census Bureau's Cartographic Boundary Files (available [here](https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html)). It file provides the polygon geometries of each county in the US; this will allow us to make visualizations of the US with breakdowns by county. In this case, we are making this processing file to allow us make a map of US with children count by county.

df_counties = gpd.read_file(DATA_DIR + '/source_files/county_shapefiles/counties.shp')
df_counties['county_fips'] = df_counties['STATEFP'] + df_counties['COUNTYFP']
df_counties = df_counties[((df_counties['STATEFP'] != '02') & (df_counties['STATEFP'] != '15') & 
                           (df_counties['STATEFP'] != '72') & (df_counties['STATEFP'] != '78') &
                           (df_counties['STATEFP'] != '60') & (df_counties['STATEFP'] != '66') &
                          (df_counties['STATEFP'] != '69'))]
df_counties.head(n=3)

### Reading InfoUSA data

This pre-processed InfoUSA household file is all the merged InfoUSA files combined in processing notebok **01_merging_files**. It contains the children counts of households as well as county FIPS of each household. This file also includes the transformed latitude and longitude coordinates.

In this dataset, the ```county_fips``` column represents each individual county; this is the column we will be grouping the children count by. After reading in the data, we filter for only the columns we will need for our visualizations.

df_hh = pd.read_parquet(DATA_DIR + '/infousa_merged.parquet')
df_hh.head()

df_hh = df_hh[['county_fips', 'state', 'child_num', 'has_child', 'age_code']]
df_hh.head()

### Using ```.groupby()``` to find the number of children per county

Here, we are taking the dataframe from above and calculating the number of children per county. The ```.groupby()``` method groups the ```df_hh``` by ```county_fips```, and the ```.sum()``` method sums the values in ```child_num``` in each ```county_fips``` group. The resulting dataframe, ```df_child_count```, has the number of children in each county.

df_child_count = df_hh.groupby('county_fips')['child_num'].sum().reset_index()
df_child_count.head()

### Using ```.merge()``` to get processed dataframe

In order to map this dataframe visually, we must have a geometry for each county. To do this, we merge ```df_child_count``` from above with ```df_counties``` ```on``` the ```county_fips``` column. Afterwards, we will select only the columns we want to keep in this dataframe and rename column names for standardization purposes before exporting this as a parquet file to be used in future visualizations.

df = df_counties.merge(df_child_count, on = ['county_fips'], how = 'left')
df.head(n=3)

df = df[['STATEFP', 'NAME', 'county_fips','child_num', 'geometry']]
df.head()

df.rename(columns = {'STATEFP': 'state', 'NAME': 'county'}, inplace = True)

df.to_file(DATA_DIR + '/children_count_by_county.shp')
